Admin, dashboard, gates, edits en soft deletes toegevoegd

+++ app/Http/Controllers/LikeController.php
@@ -1,38 +1,19 @@
 <?php
 
 namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-use App\Post;
 use App\Like;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class LikeController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index(){
 
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store($post_id)
     {
+
+        if(Gate::denies('logged-in')){
+            return redirect()->back();
+        }
 
         $like = new Like;
         $like -> user_id = Auth::id();
@@ -44,27 +25,14 @@ class LikeController extends Controller
 
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        $post = Post::find($id);
-        return view('posts.show')->withPost($post);
-    }
 
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function destroy($post_id)
     {
+
+        if(Gate::denies('logged-in')){
+            return redirect()->back();
+        }
+
         $like = Like::where('post_id', $post_id)->where('user_id', Auth::id());
         $like->delete();
