+++ app/Http/Controllers/PostController.php
@@ -2,21 +2,19 @@
 
 namespace App\Http\Controllers;
 
+
 use Illuminate\Http\Request;
 use App\Post;
 use App\User;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function index()
     {
-        $posts = Post::orderBy('created_at', 'desc')->paginate(20);
+        $posts = Post::orderBy('id', 'desc')->paginate(20);
         $users = User::all();
         return view('posts.index')->withPosts($posts)->withUsers($users);
     }
@@ -24,28 +22,39 @@ class PostController extends Controller
     public function profile($id){
 
         $user = User::find($id);
-        $posts = $user->posts;
-        return view('posts.profile', ['posts' => $posts, 'user' => $user]);
+        if($user != null) {
+            $posts = $user->posts;
+            return view('posts.profile', ['posts' => $posts, 'user' => $user]);
+        }else{
+            return redirect()->back();
+        }
+    }
+    public function dashboard(){
+
+        if(Gate::denies('admin')){
+            return redirect()->back();
+        }
+
+        $posts = Post::withTrashed()->orderBy('id', 'desc')->get();
+        $users = User::all();
+        return view('posts.dashboard')->withPosts($posts)->withUsers($users);;
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function create()
     {
+        if(Gate::denies('logged-in')){
+            return redirect()->back();
+        }
         return view('posts.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+
     public function store(Request $request)
     {
+        if(Gate::denies('logged-in')){
+            return redirect()->back();
+        }
         $this->validate($request,array(
            'description' => 'required|max:280',
             'body' => 'required'
@@ -62,54 +71,77 @@ class PostController extends Controller
 
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
         $post = Post::find($id);
-        return view('posts.show')->withPost($post);
+        if($post != null) {
+            return view('posts.show')->withPost($post);
+        }else{
+            return redirect()->back();
+        }
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function edit($id)
     {
-        //
+        $post = Post::find($id);
+        if(Gate::denies('change-post',$post)){
+            return redirect()->back();
+        }
+        if($post != null) {
+            return view('posts.edit')->withPost($post);
+        }else{
+            return redirect()->back();
+        }
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(Request $request, $id)
     {
-        //
+
+        $this->validate($request,array(
+            'description' => 'required|max:280',
+            'body' => 'required'
+        ));
+        $post = Post::find($id);
+        if(Gate::denies('change-post',$post)){
+            return redirect()->back();
+        }
+        $post ->description = $request->description;
+        $post ->body = $request->body;
+
+        $post -> save();
+
+        return redirect()->route('posts.show',$post->id);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+    public function delete($id)
+    {
+
+        $post = Post::find($id);
+        if(Gate::denies('change-post',$post)){
+            return redirect()->back();
+        }
+        if($post != null) {
+            $post->delete();
+        }
+
+        return redirect()->route('posts.profile', $post->user_id);
+    }
     public function destroy($id)
     {
-        $post = Post::find($id);
-        $post->likes()->delete();
-        //$post->shares()->delete();
-        $post->delete();
 
-        return redirect()->route('posts.profile',Auth::id());
+        if(Gate::denies('admin')){
+            return redirect()->back();
+        }
+
+        $post = Post::find($id);
+        if($post != null) {
+            $post->likes()->delete();
+            $post->shares()->delete();
+            $post->forceDelete();
+        }
+
+        return redirect()->route('posts.dashboard');
     }
 }
